refactor(ui): load JSON using fs
